+++ tests/codeforces.rs
@@ -3,7 +3,7 @@ mod service;
 use snowchains::app::{App, Opt};
 use snowchains::errors::{ServiceError, ServiceErrorKind};
 use snowchains::service::ServiceKind;
-use snowchains::terminal::{AnsiColorChoice, TermImpl};
+use snowchains::terminal::{AnsiColorChoice, Term, TermImpl};
 
 use failure::Fallible;
 use if_chain::if_chain;
@@ -57,6 +57,95 @@ fn it_fails_to_submit_if_the_lang_id_is_invalid() -> Fallible<()> {
                     Err(err.into())
                 }
             }
+        },
+    )
+}
+
+#[test]
+fn it_list_language_ids() -> Fallible<()> {
+    let _ = env_logger::try_init();
+    service::test_in_tempdir(
+        "it_list_language_ids",
+        &credentials_as_input()?,
+        |mut app| -> Fallible<()> {
+            app.run(Opt::ListLangs {
+                service: Some(ServiceKind::Codeforces),
+                contest: Some("1000".to_owned()),
+                color_choice: AnsiColorChoice::Never,
+                problem: Some("a".to_owned()),
+            })?;
+            let stdout = String::from_utf8(app.term.stdout().get_ref().to_owned())?;
+            let stderr = String::from_utf8(app.term.stderr().get_ref().to_owned())?;
+            assert_eq!(
+                stdout,
+                r#"Target: 1000/A
+GET https://codeforces.com/enter ... 200 OK
+POST https://codeforces.com/enter ... 302 Found
+GET https://codeforces.com/enter ... 302 Found
+GET https://codeforces.com/contest/1000/submit ... 200 OK
++--------------------------------+----+
+| Name                           | ID |
++--------------------------------+----+
+| GNU GCC C11 5.1.0              | 43 |
++--------------------------------+----+
+| Clang++17 Diagnostics          | 52 |
++--------------------------------+----+
+| GNU G++11 5.1.0                | 42 |
++--------------------------------+----+
+| GNU G++14 6.4.0                | 50 |
++--------------------------------+----+
+| GNU G++17 7.3.0                | 54 |
++--------------------------------+----+
+| Microsoft Visual C++ 2010      | 2  |
++--------------------------------+----+
+| Microsoft Visual C++ 2017      | 59 |
++--------------------------------+----+
+| C# Mono 5.18                   | 9  |
++--------------------------------+----+
+| D DMD32 v2.083.1               | 28 |
++--------------------------------+----+
+| Go 1.11.4                      | 32 |
++--------------------------------+----+
+| Haskell GHC 7.8.3 (2014.2.0.0) | 12 |
++--------------------------------+----+
+| Java 1.8.0_162                 | 36 |
++--------------------------------+----+
+| Kotlin 1.3.10                  | 48 |
++--------------------------------+----+
+| OCaml 4.02.1                   | 19 |
++--------------------------------+----+
+| Delphi 7                       | 3  |
++--------------------------------+----+
+| Free Pascal 3.0.2              | 4  |
++--------------------------------+----+
+| PascalABC.NET 3.4.2            | 51 |
++--------------------------------+----+
+| Perl 5.20.1                    | 13 |
++--------------------------------+----+
+| PHP 7.2.13                     | 6  |
++--------------------------------+----+
+| Python 2.7.15                  | 7  |
++--------------------------------+----+
+| Python 3.7.2                   | 31 |
++--------------------------------+----+
+| PyPy 2.7 (6.0.0)               | 40 |
++--------------------------------+----+
+| PyPy 3.5 (6.0.0)               | 41 |
++--------------------------------+----+
+| Ruby 2.0.0p645                 | 8  |
++--------------------------------+----+
+| Rust 1.31.1                    | 49 |
++--------------------------------+----+
+| Scala 2.12.8                   | 20 |
++--------------------------------+----+
+| JavaScript V8 4.8.0            | 34 |
++--------------------------------+----+
+| Node.js 9.4.0                  | 55 |
++--------------------------------+----+
+"#,
+            );
+            assert_eq!(stderr, "Handle/Email: Password: ");
+            Ok(())
         },
     )
 }